+++ src/config/local.sample.js
@@ -1,12 +1,6 @@
 export default {
   api: {
-    // Prod
-    // endpoint: '//api.gw2armory.com/',
-
-    // Boot2Docker
-    // endpoint: '//192.168.59.103/',
-
-    // Docker for Mac/Windows
+    endpoint: '//api.gw2armory.com/',
     // endpoint: '//localhost/',
   },
 };