+++ src/features/User/components/DailyAp/index.js
@@ -2,23 +2,27 @@ import { PropTypes } from 'react';
 
 import Summary from 'common/layouts/Summary';
 import ProgressBar from 'common/components/ProgressBar';
-import Icon from 'common/components/Icon';
 import T from 'i18n-react';
+import Redacted from 'common/components/Redacted';
 
-const DailyAp = ({ dailyAp, monthlyAp }) => (
-  <Summary
-    leftIcon={<Icon name="daily.png" size="xlarge" />}
-    title={T.translate('accSummary.dailyAp')}
-    subTitle={
-      <ProgressBar
-        current={dailyAp + monthlyAp}
-        max={15000}
-        backgroundColor="rgb(41, 41, 41)"
-        barColor="rgb(85, 35, 164)"
-      />
-    }
-  />
-);
+const DailyAp = ({ dailyAp, monthlyAp }) => {
+  const count = (dailyAp + monthlyAp) || 0;
+
+  return (
+    <Summary
+      leftIcon={{ name: 'daily.png', size: 'xlarge' }}
+      title={<Redacted redact={!count}>{T.translate('accSummary.dailyAp')}</Redacted>}
+      subTitle={
+        <ProgressBar
+          current={count}
+          max={15000}
+          backgroundColor="rgb(41, 41, 41)"
+          barColor="rgb(85, 35, 164)"
+        />
+      }
+    />
+  );
+};
 
 DailyAp.propTypes = {
   dailyAp: PropTypes.number,