+++ src/features/User/components/RaidSummary/index.js
@@ -1,9 +1,9 @@
 import { PropTypes } from 'react';
 
 import ProgressBar from 'common/components/ProgressBar';
-import Icon from 'common/components/Icon';
 import Summary from 'common/layouts/Summary';
 import T from 'i18n-react';
+import Redacted from 'common/components/Redacted';
 
 const RAID_KILL_ACHIEVEMENTS = {
   2654: 'Vg', // Vale Guardian
@@ -18,11 +18,12 @@ const RAID_KILL_ACHIEVEMENTS = {
 const TOTAL_KILLS = Object.keys(RAID_KILL_ACHIEVEMENTS).length;
 
 const RaidSummary = ({ userAchievements, className, simple }) => {
+  const redact = !userAchievements.length;
   const achievements = userAchievements.filter(
     ({ id, done }) => RAID_KILL_ACHIEVEMENTS[id] && done
   );
 
-  const raidKills = T.translate('accSummary.raidBossKills') + `${achievements.map(({ id }) =>
+  const raidKills = `${T.translate('accSummary.raidBossKills')} ${achievements.map(({ id }) =>
     `[${RAID_KILL_ACHIEVEMENTS[id]}]`).join('')}`;
 
   if (simple) {
@@ -32,8 +33,8 @@ const RaidSummary = ({ userAchievements, className, simple }) => {
   return (
     <Summary
       className={className}
-      leftIcon={<Icon name="raid.png" size="xlarge" />}
-      title={raidKills}
+      leftIcon={{ name: 'raid.png', size: 'xlarge' }}
+      title={<Redacted redact={redact}>{raidKills}</Redacted>}
       subTitle={
         <ProgressBar
           current={achievements.length}